finished redoing the server
