+++ server/config/config.js
@@ -1,5 +1,5 @@
 // sets secrets to either be the heroku variable or, if not accessing through heroku, will be set through a file which is on the project but not pushed to github or heroku.
-var secrets = process.env.PASSCODE === undefined ? require('./secretpass.js') : 'test';
+var secrets = process.env.PASSCODE === undefined ? require('../secretpass.js') : 'test';
 var passcode = secrets.passcode;
 
 module.exports ={
